requirements.txt
python-binance==1.0.19
flask==3.1.0
flask-socketio==5.3.6
requests==2.31.0
websocket-client==1.6.4